requirements.txt
fastapi
uvicorn
python-dotenv
openai
faiss-cpu
streamlit
plotly
sqlalchemy
requests